Fix/separate/improve tests for retriever and api client
